move code to the src directory
I think this makes it easier to maintain and is more consistent with other typescript projects
