+++ src/DomainConfig.ts
