+++ src/DomainInfo.ts
