+++ src/Globals.ts
