+++ tsconfig.json
@@ -2,17 +2,13 @@
   "compilerOptions": {
     "module": "commonjs",
     "moduleResolution": "node",
-    "rootDir": ".",
+    "rootDir": "src",
     "target": "es6",
     "sourceMap": false,
     "outDir": "dist",
     "types": ["node"]
   },
   "include": [
-    "**/*.ts"
-  ],
-  "exclude": [
-    "node_modules",
-    "test"
+    "src/**/*.ts"
   ]
 }
