+++ tsconfig.tests.json
@@ -9,9 +9,7 @@
     "types": ["mocha", "node"]
   },
   "include": [
+    "src/**/*.ts",
     "test/**/*.ts"
-  ],
-  "exclude": [
-    "node_modules"
   ]
 }
